Fix usage example
The usage example did not previously compile because it used
`.description`, which didn't exist.  The commit fixes the code to use
`.help`, which does exists.

It also prints the output rather than saving it to an unused variable,
which allows the `preview` instructions below the usage example to
function as described.