Add custom section (#24)
* Add custom section method

This commit adds a `.custom` method to the public API, which takes a
section title and 0 or more paragraphs of plain text.  Using this
method, the following code
```
.custom(
    Sec::new("regex syntax")
        .paragraph("ripgrep uses Rust’s regex engine by default, which documents its syntax: https://docs.rs/regex/*/regex/#syntax")
)
```
could generate the first portion of `ripgrep`'s REGEX SYNTAX section.

* Add documentation for custom section

* Changes suggested by @yoshuawuyts

This commit changes the section struct from `Sec` to `Section` to avoid
ambiguity with "security".  It also adds a more descriptive doc comment
to section.rs

* Update README and example with new Section struct