Update README example with bold
This commit updates the README example output to show that `man` outputs
some text in bold.  It does so by switching from GitHub-style codeblocks
to the `<pre>` html tag, which allows limited inline styling.

It uses the `<b>` html tag rather than the (more modern) `<strong>` tag
because the `<b>` tag is preferred in those limited situations where the
meaning is that the text is rendered in bold rather than that the text
deserves strong emphasis.

Additionally, it uses the `<i>` tag to italicise the argument passed to
options even though this text is underlined in the generated man page.
This is because GitHub does not render inline `<span
style="text-decoration: underline">` and I thought indicating that this
text is styled a bit differently would be better than leaving it as
regular unstyled roman text.